# SimpleSocialNetwork/client_api_gen/swagger_api.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SimpleSocialNetworkBack",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Subscription/subscribed_to": {
      "get": {
        "tags": [
          "Subscription"
        ],
        "summary": "Get all users the specified user is subscribed to",
        "parameters": [
          {
            "name": "login",
            "in": "query",
            "description": "Login of a user",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Login of a user"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubscriptionModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Subscription/is_subscribed_to": {
      "get": {
        "tags": [
          "Subscription"
        ],
        "summary": "Check whether current user is subscribed to the specified user",
        "parameters": [
          {
            "name": "login",
            "in": "query",
            "description": "User login",
            "required": true,
            "schema": {
              "type": "string",
              "description": "User login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/Subscription/sub": {
      "post": {
        "tags": [
          "Subscription"
        ],
        "summary": "Subscribe current user to the specified user",
        "parameters": [
          {
            "name": "login",
            "in": "query",
            "description": "Specified user's login",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Specified user's login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Subscription/unsub": {
      "post": {
        "tags": [
          "Subscription"
        ],
        "summary": "Unsubscribe current user from the specified user",
        "parameters": [
          {
            "name": "login",
            "in": "query",
            "description": "Specified user's login",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Specified user's login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/exists": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Check whether the specified user exists (deleted users also regarded as existing)",
        "parameters": [
          {
            "name": "login",
            "in": "query",
            "description": "The specified user's login",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The specified user's login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/deleted": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Check whether the specified user was deleted",
        "parameters": [
          {
            "name": "login",
            "in": "query",
            "description": "The specified user's login",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The specified user's login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/User": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get a full model of the specified user",
        "parameters": [
          {
            "name": "login",
            "in": "query",
            "description": "Login of a user",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Login of a user"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserModel"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Soft delete user;\r\nAuthorized: either user themselves or an admin",
        "parameters": [
          {
            "name": "login",
            "in": "query",
            "description": "Login of a user",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Login of a user"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/info": {
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Change user info",
        "requestBody": {
          "description": "New user info",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeUserInfo"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeUserInfo"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeUserInfo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/limited": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get a limited model of the specified user",
        "parameters": [
          {
            "name": "login",
            "in": "query",
            "description": "Login of a user",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Login of a user"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LimitedUserModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/search": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Search users",
        "parameters": [
          {
            "name": "NamePattern",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "AboutPattern",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LimitedUserModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/User/elevate": {
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Elevate the specified user to admin",
        "parameters": [
          {
            "name": "login",
            "in": "query",
            "description": "Login of a user",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Login of a user"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/ban": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Ban the specified user",
        "requestBody": {
          "description": "Ban data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BanUserModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BanUserModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BanUserModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/unban": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Lift all current bans of the specified user",
        "parameters": [
          {
            "name": "user",
            "in": "query",
            "description": "Login of a user",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Login of a user"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/banned": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Check whether user is banned",
        "parameters": [
          {
            "name": "login",
            "in": "query",
            "description": "Login of a user",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Login of a user"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "UserModel": {
        "required": [
          "isAdmin",
          "isDeleted",
          "login"
        ],
        "type": "object",
        "properties": {
          "login": {
            "type": "string"
          },
          "about": {
            "type": "string",
            "nullable": true
          },
          "dateBirth": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "isAdmin": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "SubscriptionModel": {
        "required": [
          "id",
          "isActive"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "subscriberId": {
            "type": "string",
            "nullable": true
          },
          "targetId": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ChangeUserInfo": {
        "type": "object",
        "properties": {
          "about": {
            "type": "string",
            "nullable": true
          },
          "dateBirth": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LimitedUserModel": {
        "required": [
          "about",
          "isDeleted",
          "login"
        ],
        "type": "object",
        "properties": {
          "login": {
            "type": "string"
          },
          "about": {
            "type": "string"
          },
          "isDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "BanUserModel": {
        "required": [
          "expirationDate",
          "login"
        ],
        "type": "object",
        "properties": {
          "login": {
            "type": "string"
          },
          "expirationDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      }
    }
  }
}